terminando empleado

+++ PetShop/EntidadesPetShop/Empleado.cs
@@ -42,7 +42,7 @@ namespace EntidadesPetShop
             ultimoIdGenerado = 99999;
         }
 
-        public Empleado(string nombre, string apellido, string dni, long cuil, Usuario usuario, double sueldo) : base(nombre, apellido, dni, cuil)
+        public Empleado(string nombre, string apellido, string dni, double cuil, Usuario usuario, double sueldo) : base(nombre, apellido, dni, cuil)
         {
             ultimoIdGenerado++;
             this.Id = ultimoIdGenerado;
